Move config_yaml.h to public tpcrs/ include path
