+++ src/CMakeLists.txt
@@ -2,6 +2,7 @@ configure_file(config.h.in config.h @ONLY)
 
 set(header_files
     ${TPCRS_SOURCE_DIR}/include/tpcrs/configurator.h
+    ${TPCRS_SOURCE_DIR}/include/tpcrs/config_yaml.h
     ${TPCRS_SOURCE_DIR}/include/tpcrs/structs.h
     ${TPCRS_SOURCE_DIR}/include/tpcrs/tpcrs.h
 )
